requirements.txt
uvicorn
python-dotenv
fastapi
pydantic
openai
